LayoutTests/http/tests/security/frameNavigation/navigation-DENIED-sibling-location-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to initiate a navigation change for frame with URL http://127.0.0.1:8000/security/frameNavigation/resources/ready.html from frame with URL http://localhost:8000/security/frameNavigation/resources/ready.html.

Navigating frame.
 

--------
Frame: 'child0'
--------
127.0.0.1:8000

Waiting for navigation commands.

--------
Frame: 'active0'
--------
localhost:8000

Waiting for navigation commands.
